refactor: separate view model from view (starting screen) Also made this screen's viewmodel architecture cleaner
