+++ src/main/kotlin/model/functionality/iograph/GraphType.kt
@@ -1,8 +1,8 @@
 package model.functionality.iograph
 
-enum class GraphType(val type: String) {
-    UNDIRECTED_GRAPH("UndirectedGraph"),
-    DIRECTED_GRAPH("DirectedGraph"),
-    UNDIRECTED_WEIGHTED_GRAPH("UndirectedWeightedGraph"),
-    DIRECTED_WEIGHTED_GRAPH("DirectedWeightedGraph"),
+enum class GraphType(val string: String) {
+    UNDIRECTED_GRAPH("Undirected Graph"),
+    DIRECTED_GRAPH("Directed Graph"),
+    UNDIRECTED_WEIGHTED_GRAPH("Undirected Weighted Graph"),
+    DIRECTED_WEIGHTED_GRAPH("Directed Weighted Graph"),
 }
